Move ADE, ADM, AEBN and Popporn classes to nzedb/processing/adult
